# config/banking-api-openapi.yaml
openapi: 3.0.3
info:
  title: Banking Services API
  version: 1.0.0
  description: |
    # Banking Services API
    
    A comprehensive RESTful API for banking operations including account management, 
    transactions, payments, and beneficiary management.
    
    ## Features
    - Account management and balance inquiries
    - Transaction history and search
    - Domestic and international payments
    - Beneficiary management
    - Account statements generation
    
    ## Security
    This API uses OAuth 2.0 for authentication and authorization. All endpoints 
    require a valid access token except for the health check endpoint.
    
    ## Rate Limiting
    Rate limits are enforced based on your subscription tier:
    - Bronze: 10 requests per minute
    - Silver: 100 requests per minute
    - Gold: 1000 requests per minute
    
    ## Support
    For API support, contact: api-support@bankingservices.com
  contact:
    name: API Support Team
    email: api-support@bankingservices.com
    url: https://developer.bankingservices.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.bankingservices.com/terms
servers:
  - url: https://api.bankingservices.com/v1
    description: Production server
  - url: https://sandbox-api.bankingservices.com/v1
    description: Sandbox server for testing

# API Connect v10 Configuration
...
security:
  - oauth2ClientCredentials: []
  - oauth2AuthCode: []
  - apiKey: []
...
paths:
...
  /payments:
    post:
      tags:
        - Payments
      summary: Create a payment
      description: Initiate a new payment transaction (domestic or international)
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            example:
              fromAccountId: ACC-123456789
              toBeneficiaryId: BEN-987654321
              amount: 500.00
              currency: USD
              paymentType: DOMESTIC
              reference: Invoice #12345
              scheduledDate: '2026-01-10'
              urgency: STANDARD
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              example:
                paymentId: PAY-20260109-001
                status: PENDING
                fromAccountId: ACC-123456789
                toBeneficiaryId: BEN-987654321
                amount: 500.00
                currency: USD
                paymentType: DOMESTIC
                reference: Invoice #12345
                scheduledDate: '2026-01-10'
                createdAt: '2026-01-09T16:45:00Z'
                estimatedCompletionDate: '2026-01-10T12:00:00Z'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payments/{paymentId}:
    get:
      tags:
        - Payments
      summary: Get payment status
      description: Retrieve the current status and details of a payment
      operationId: getPaymentById
      parameters:
        - $ref: '#/components/parameters/PaymentIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              example:
                paymentId: PAY-20260109-001
                status: COMPLETED
                fromAccountId: ACC-123456789
                toBeneficiaryId: BEN-987654321
                amount: 500.00
                currency: USD
                paymentType: DOMESTIC
                reference: Invoice #12345
                scheduledDate: '2026-01-10'
                createdAt: '2026-01-09T16:45:00Z'
                completedAt: '2026-01-10T09:30:00Z'
                transactionId: TXN-20260110-023
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payments/{paymentId}/cancel:
    put:
      tags:
        - Payments
      summary: Cancel a payment
      description: Cancel a pending or scheduled payment
      operationId: cancelPayment
      parameters:
        - $ref: '#/components/parameters/PaymentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  type: string
                  description: Reason for cancellation
                  minLength: 10
                  maxLength: 500
            example:
              reason: Payment no longer required due to invoice cancellation
      responses:
        '200':
          description: Payment cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              example:
                paymentId: PAY-20260109-001
                status: CANCELLED
                fromAccountId: ACC-123456789
                toBeneficiaryId: BEN-987654321
                amount: 500.00
                currency: USD
                paymentType: DOMESTIC
                reference: Invoice #12345
                scheduledDate: '2026-01-10'
                createdAt: '2026-01-09T16:45:00Z'
                cancelledAt: '2026-01-09T17:00:00Z'
                cancellationReason: Payment no longer required due to invoice cancellation
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /beneficiaries:
    get:
      tags:
        - Beneficiaries
      summary: List all beneficiaries
      description: Retrieve a list of all registered beneficiaries
      operationId: getBeneficiaries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: beneficiaryType
          in: query
          description: Filter by beneficiary type
          schema:
            type: string
            enum: [INDIVIDUAL, BUSINESS]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryListResponse'
              example:
                beneficiaries:
                  - beneficiaryId: BEN-987654321
                    beneficiaryType: INDIVIDUAL
                    name: John Smith
                    nickname: John - Rent
                    accountNumber: '****9876'
                    bankName: First National Bank
                    status: ACTIVE
                    createdAt: '2025-06-15T10:00:00Z'
                  - beneficiaryId: BEN-123456789
                    beneficiaryType: BUSINESS
                    name: Acme Corporation
                    nickname: Acme - Supplies
                    accountNumber: '****5432'
                    bankName: Business Bank
                    status: ACTIVE
                    createdAt: '2025-08-20T14:30:00Z'
                pagination:
                  page: 1
                  limit: 10
                  totalPages: 1
                  totalRecords: 2
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Beneficiaries
      summary: Add a beneficiary
      description: Register a new beneficiary for payments
      operationId: createBeneficiary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeneficiaryRequest'
            example:
              beneficiaryType: INDIVIDUAL
              name: Jane Doe
              nickname: Jane - Utilities
              accountNumber: '1234567890123456'
              routingNumber: '021000021'
              bankName: Second National Bank
              bankAddress:
                street: 456 Bank Street
                city: Boston
                state: MA
                postalCode: '02101'
                country: US
              email: jane.doe@example.com
              phone: '+1-555-0123'
      responses:
        '201':
          description: Beneficiary created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryDetails'
              example:
                beneficiaryId: BEN-456789123
                beneficiaryType: INDIVIDUAL
                name: Jane Doe
                nickname: Jane - Utilities
                accountNumber: '****3456'
                routingNumber: '021000021'
                bankName: Second National Bank
                bankAddress:
                  street: 456 Bank Street
                  city: Boston
                  state: MA
                  postalCode: '02101'
                  country: US
                email: jane.doe@example.com
                phone: '+1-555-0123'
                status: PENDING_VERIFICATION
                createdAt: '2026-01-09T16:45:00Z'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /beneficiaries/{beneficiaryId}:
    get:
      tags:
        - Beneficiaries
      summary: Get beneficiary details
      description: Retrieve detailed information for a specific beneficiary
      operationId: getBeneficiaryById
      parameters:
        - $ref: '#/components/parameters/BeneficiaryIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryDetails'
              example:
                beneficiaryId: BEN-987654321
                beneficiaryType: INDIVIDUAL
                name: John Smith
                nickname: John - Rent
                accountNumber: '****9876'
                routingNumber: '021000022'
                bankName: First National Bank
                bankAddress:
                  street: 123 Main Street
                  city: New York
                  state: NY
                  postalCode: '10001'
                  country: US
                email: john.smith@example.com
                phone: '+1-555-0124'
                status: ACTIVE
                createdAt: '2025-06-15T10:00:00Z'
                lastUsed: '2026-01-05T14:20:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Beneficiaries
      summary: Update beneficiary
      description: Update beneficiary information
      operationId: updateBeneficiary
      parameters:
        - $ref: '#/components/parameters/BeneficiaryIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeneficiaryUpdateRequest'
            example:
              nickname: John - Monthly Rent
              email: john.smith.new@example.com
              phone: '+1-555-9999'
      responses:
        '200':
          description: Beneficiary updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeneficiaryDetails'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Beneficiaries
      summary: Delete beneficiary
      description: Remove a beneficiary from the registered list
      operationId: deleteBeneficiary
      parameters:
        - $ref: '#/components/parameters/BeneficiaryIdParam'
      responses:
        '204':
          description: Beneficiary deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          description: Cannot delete beneficiary with pending payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: BENEFICIARY_IN_USE
                  message: Cannot delete beneficiary with pending payments
                  timestamp: '2026-01-09T16:45:00Z'
                  correlationId: corr-123456
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    oauth2ClientCredentials:
      type: oauth2
      description: OAuth 2.0 Client Credentials flow for server-to-server authentication
      flows:
        clientCredentials:
          tokenUrl: https://auth.bankingservices.com/oauth2/token
          scopes:
            accounts:read: Read account information
            accounts:write: Modify account information
            transactions:read: Read transaction history
            payments:read: Read payment information
            payments:write: Create and manage payments
            beneficiaries:read: Read beneficiary information
            beneficiaries:write: Manage beneficiaries
            statements:read: Access account statements
      x-ibm-configuration:
        provider: oauth-provider
        
    oauth2AuthCode:
      type: oauth2
      description: OAuth 2.0 Authorization Code flow for user authentication
      flows:
        authorizationCode:
          authorizationUrl: https://auth.bankingservices.com/oauth2/authorize
          tokenUrl: https://auth.bankingservices.com/oauth2/token
          scopes:
            accounts:read: Read account information
            accounts:write: Modify account information
            transactions:read: Read transaction history
            payments:read: Read payment information
            payments:write: Create and manage payments
            beneficiaries:read: Read beneficiary information
            beneficiaries:write: Manage beneficiaries
            statements:read: Access account statements
      x-ibm-configuration:
        provider: oauth-provider
        
    apiKey:
      type: apiKey
      description: API key for application identification
...
  parameters:
...
    PaymentIdParam:
      name: paymentId
      in: path
      required: true
      description: Unique identifier for the payment
      schema:
        type: string
        pattern: '^PAY-[0-9]{8}-[0-9]{3}$'
      example: PAY-20260109-001
    BeneficiaryIdParam:
      name: beneficiaryId
      in: path
      required: true
      description: Unique identifier for the beneficiary
      schema:
        type: string
        pattern: '^BEN-[0-9]{9}$'
      example: BEN-987654321
...
    PageParam:
      name: page
      in: query
      description: Page number for pagination (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
  schemas:
...
    PaymentRequest:
      type: object
      required:
        - fromAccountId
        - toBeneficiaryId
        - amount
        - currency
        - paymentType
      properties:
        fromAccountId:
          type: string
          description: Source account identifier
        toBeneficiaryId:
          type: string
          description: Beneficiary identifier
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Payment amount
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code
        paymentType:
          type: string
          enum: [DOMESTIC, INTERNATIONAL, WIRE]
          description: Type of payment
        reference:
          type: string
          maxLength: 140
          description: Payment reference or description
        scheduledDate:
          type: string
          format: date
          description: Scheduled payment date (defaults to today)
        urgency:
          type: string
          enum: [STANDARD, URGENT, EXPRESS]
          default: STANDARD
          description: Payment urgency level
    PaymentResponse:
      type: object
      required:
        - paymentId
        - status
        - fromAccountId
        - toBeneficiaryId
        - amount
        - currency
        - paymentType
        - createdAt
      properties:
        paymentId:
          type: string
          description: Unique payment identifier
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED]
          description: Payment status
        fromAccountId:
          type: string
          description: Source account identifier
        toBeneficiaryId:
          type: string
          description: Beneficiary identifier
        amount:
          type: number
          format: double
          description: Payment amount
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code
        paymentType:
          type: string
          enum: [DOMESTIC, INTERNATIONAL, WIRE]
          description: Type of payment
        reference:
          type: string
          description: Payment reference
        scheduledDate:
          type: string
          format: date
          description: Scheduled payment date
        createdAt:
          type: string
          format: date-time
          description: Payment creation timestamp
        completedAt:
          type: string
          format: date-time
          description: Payment completion timestamp
        cancelledAt:
          type: string
          format: date-time
          description: Payment cancellation timestamp
        estimatedCompletionDate:
          type: string
          format: date-time
          description: Estimated completion date
        transactionId:
          type: string
          description: Associated transaction ID (when completed)
        cancellationReason:
          type: string
          description: Reason for cancellation
    BeneficiaryListResponse:
      type: object
      required:
        - beneficiaries
        - pagination
      properties:
        beneficiaries:
          type: array
          items:
            $ref: '#/components/schemas/BeneficiarySummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    BeneficiaryRequest:
      type: object
      required:
        - beneficiaryType
        - name
        - accountNumber
        - routingNumber
        - bankName
      properties:
        beneficiaryType:
          type: string
          enum: [INDIVIDUAL, BUSINESS]
          description: Type of beneficiary
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Beneficiary full name
        nickname:
          type: string
          maxLength: 50
          description: User-defined nickname
        accountNumber:
          type: string
          pattern: '^[0-9]{10,18}$'
          description: Beneficiary account number
        routingNumber:
          type: string
          pattern: '^[0-9]{9}$'
          description: Bank routing number
        bankName:
          type: string
          description: Bank name
        bankAddress:
          $ref: '#/components/schemas/Address'
        email:
          type: string
          format: email
          description: Beneficiary email address
        phone:
          type: string
          pattern: '^\+[1-9]\d{1,14}$'
          description: Phone number in E.164 format
    BeneficiaryDetails:
      allOf:
        - $ref: '#/components/schemas/BeneficiarySummary'
        - type: object
          properties:
            routingNumber:
              type: string
              description: Bank routing number
            bankAddress:
              $ref: '#/components/schemas/Address'
            email:
              type: string
              format: email
              description: Email address
            phone:
              type: string
              description: Phone number
            lastUsed:
              type: string
              format: date-time
              description: Last payment date
    BeneficiaryUpdateRequest:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 50
          description: User-defined nickname
        email:
          type: string
          format: email
          description: Email address
        phone:
          type: string
          pattern: '^\+[1-9]\d{1,14}$'
          description: Phone number in E.164 format
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              description: Additional error details
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
            correlationId:
              type: string
              description: Request correlation ID for tracking
  responses:
    BadRequestError:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: Invalid request parameters
              details:
                - field: amount
                  message: Amount must be greater than 0
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired access token
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: Insufficient permissions to access this resource
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    NotFoundError:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: The requested resource was not found
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    UnprocessableEntityError:
      description: Unprocessable entity - Business logic validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNPROCESSABLE_ENTITY
              message: Payment cannot be processed
              details:
                - field: amount
                  message: Insufficient funds in account
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    RateLimitError:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when the rate limit resets (Unix timestamp)
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Rate limit exceeded. Please try again later.
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred. Please try again later.
              timestamp: '2026-01-09T16:45:00Z'
              correlationId: corr-123456

# Made with Bob